+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+## [2.13.0] - 2020-11-27
 ### Added
 - Add 'copy citation' button to artwork detail